judgeflow_backend::utils::csv_processing

Function read_csv_records

pub fn read_csv_records<T, P>(path: P) -> Result<Vec<T>, Box<dyn Error>>
where
    T: DeserializeOwned,
    P: AsRef<Path>,

Reads records from a CSV file into a vector of the specified type.

This is a generic function that can parse any CSV file into a vector of any type that implements serde::de::DeserializeOwned.

Arguments

path - Path to the CSV file to read

Returns

A Result containing a vector of deserialized records on success, or an error if reading or parsing fails.

Errors

Returns an error if:

The file cannot be opened

The CSV is malformed

The data cannot be deserialized into the target type

Example

use std::path::Path;
use serde::Deserialize;
use judgeflow_backend::utils::csv_processing::read_csv_records;

#[derive(Debug, Deserialize)]
struct Person {
    name: String,
    age: u32,
    email: String,
}

let people: Vec<Person> = read_csv_records("data/people.csv")?;
for person in people {
    println!("{} ({}): {}", person.name, person.age, person.email);
}